Restore __main__ module after loading and running the plugin.
If the plugin raises an exception (including SystemExit) the __main__ module is
not properly loaded and it's __dict__ is lost. E.g. this causes errors when
using any global variable in planet.py. I don't think is is a good idea to keep
the plugin in memory (as __main__) anyway.